requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.2.1
email-validator==2.2.0
Werkzeug==3.0.3
SQLAlchemy==2.0.36
python-dotenv==1.0.1
gunicorn==21.2.0
